Refactoring paths: all frontend JS files are into './src/public/js/' dir
