+++ src/public/index.html
@@ -25,6 +25,6 @@
     <p id="game-status">Bienvenido</p>
   </footer>
   
-  <script type="module" src="./ui.js"></script>
+  <script type="module" src="./js/ui.js"></script>
 </body>
 </html>
